# stylo: Improve restyling performance