--- a/components/constellation/constellation.rs
+++ b/components/constellation/constellation.rs
@@ -1441,6 +1441,7 @@ impl<Message, LTF, STF> Constellation<Message, LTF, STF>
             let root = self.root_frame_id.is_none() || self.root_frame_id == Some(frame_id);
             self.compositor_proxy.send(ToCompositorMsg::LoadComplete(back, forward, root));
         }
+        self.handle_subframe_loaded(pipeline_id);
     }
 
     fn handle_dom_load(&mut self, pipeline_id: PipelineId) {
@@ -1455,8 +1456,6 @@ impl<Message, LTF, STF> Constellation<Message, LTF, STF>
         if webdriver_reset {
             self.webdriver.load_channel = None;
         }
-
-        self.handle_subframe_loaded(pipeline_id);
     }
 
     fn handle_traverse_history_msg(&mut self,
@@ -2073,7 +2072,7 @@ impl<Message, LTF, STF> Constellation<Message, LTF, STF>
         }
 
         // If there are pending loads, wait for those to complete.
-        if self.pending_frames.len() > 0 {
+        if !self.pending_frames.is_empty() {
             return ReadyToSave::PendingFrames;
         }
 
@@ -2089,7 +2088,10 @@ impl<Message, LTF, STF> Constellation<Message, LTF, STF>
             let pipeline_id = frame.current.0;
 
             let pipeline = match self.pipelines.get(&pipeline_id) {
-                None => { warn!("Pipeline {:?} screenshot while closing.", pipeline_id); continue; },
+                None => {
+                    warn!("Pipeline {:?} screenshot while closing.", pipeline_id);
+                    continue;
+                },
                 Some(pipeline) => pipeline,
             };