--- a/components/constellation/pipeline.rs
+++ b/components/constellation/pipeline.rs
@@ -501,6 +501,10 @@ impl UnprivilegedPipelineContent {
                 command.env("RUST_BACKTRACE", value);
             }
 
+            if let Ok(value) = env::var("RUST_LOG") {
+                command.env("RUST_LOG", value);
+            }
+
             let profile = content_process_sandbox_profile();
             ChildProcess::Sandboxed(Sandbox::new(profile).start(&mut command)
                                     .expect("Failed to start sandboxed child process!"))
@@ -513,6 +517,10 @@ impl UnprivilegedPipelineContent {
 
             if let Ok(value) = env::var("RUST_BACKTRACE") {
                 child_process.env("RUST_BACKTRACE", value);
+            }
+
+            if let Ok(value) = env::var("RUST_LOG") {
+                child_process.env("RUST_LOG", value);
             }
 
             ChildProcess::Unsandboxed(child_process.spawn()