@emilio 2016-07-23
I *think* these are servo incremental layout bugs uncovered by this patch (since right now we relayout way less than before), I'll verify this during the weekend.