@emilio 2016-07-23
What a better way to know?

@bors-servo: try