@bors-servo 2016-07-23
:hourglass: Trying commit 01a8d96 with merge 37d2095...