@emilio 2016-07-23
@bors-servo: try

I think I've got a few (all?) of them, was a super-stupid thing, layout only checked for `has_dirty_descendants`, which in practice **before** it meant `is_dirty || has_dirty_descendants` due to how the flags were propagated.