@bors-servo 2016-07-23
:hourglass: Trying commit 7be7642 with merge 560629b...