@bors-servo 2016-07-23
:hourglass: Trying commit eab954a with merge a6ca1e0...