@bors-servo 2016-07-24
:hourglass: Trying commit fe56238 with merge 33abd44...