@bors-servo 2016-07-25
:sunny: Test successful - [android](http://build.servo.org/builders/android/builds/2029), [arm32](http://build.servo.org/builders/arm32/builds/2142), [arm64](http://build.servo.org/builders/arm64/builds/2144), [linux-dev](http://build.servo.org/builders/linux-dev/builds/2187), [linux-rel](http://build.servo.org/builders/linux-rel/builds/2311), [mac-dev-unit](http://build.servo.org/builders/mac-dev-unit/builds/2217), [mac-rel-css](http://build.servo.org/builders/mac-rel-css/builds/2203), [mac-rel-wpt](http://build.servo.org/builders/mac-rel-wpt/builds/2275), [windows-dev](http://build.servo.org/builders/windows-dev/builds/89)