@emilio 2016-07-25
@samlh Sure, so as an initial way to measure, I just grabbed a twitter feed page, removed scripts and iframes, and loaded it with a release build of servo before and after this patch, profiling with `-p 5`, navigating normally trying to interact with elements to trigger incremental restyle. 

Unfortunately I'm worried this is not a exact measure (even the tiniest difference in interaction could trigger way different code paths), but it might be a good approx measure if done for enough time (these are about 120 incremental restyle events, which are not that many, but probably enough).

Anyway, in that (unexact, but probably sort of relevant) test, the layout numbers are as follows:

### Before patch

```
_category_                          _incremental?_ _iframe?_             _url_                  _mean (ms)_   _median (ms)_      _min (ms)_      _max (ms)_       _events_ 
Compositing                             N/A          N/A                  N/A                        0.6851          0.1803          0.0626        196.0905             559
Layout                                  yes          yes     file:///home/emilio/projects/m         76.6177         55.2281         28.1311       5692.9606             385
Layout                                  no           yes     file:///home/emilio/projects/m        990.0932        990.0932        990.0932        990.0932               1
+ Style Recalc                          yes          yes     file:///home/emilio/projects/m         11.2728         10.4425          3.5714        101.0940             117
+ Style Recalc                          no           yes     file:///home/emilio/projects/m        840.2206        840.2206        840.2206        840.2206               1
| + Text Shaping                        yes          yes     file:///home/emilio/projects/m          0.0583          0.0046          0.0000          2.2492              93
| + Text Shaping                        no           yes     file:///home/emilio/projects/m          2.7505          2.7505          2.7505          2.7505               1
```

### After patch

```
_category_                          _incremental?_ _iframe?_             _url_                  _mean (ms)_   _median (ms)_      _min (ms)_      _max (ms)_       _events_ 
Compositing                             N/A          N/A                  N/A                        0.4669          0.1659          0.0518        131.3618             799
Layout                                  yes          yes     file:///home/emilio/projects/m         74.4345         55.6957         48.5256       6706.9546             399
Layout                                  no           yes     file:///home/emilio/projects/m       1096.8402       1096.8402       1096.8402       1096.8402               1
+ Style Recalc                          yes          yes     file:///home/emilio/projects/m          5.3830          3.6911          0.9647        182.8923             138
+ Style Recalc                          no           yes     file:///home/emilio/projects/m        840.3075        840.3075        840.3075        840.3075               1
| + Text Shaping                        yes          yes     file:///home/emilio/projects/m          0.0483          0.0053          0.0000          2.7329             115
| + Text Shaping                        no           yes     file:///home/emilio/projects/m          2.4944          2.4944          2.4944          2.4944               1
```

Which is more than a 50% improvement on mean incremental style recalc (5.3s vs 11.2s). Again, while this allows some sort of measurement, take these with a *huge* grain of salt, since these are not exact tests. Also, note that this matters even more for the case I was optimizing (stylo), since there we avoid way more FFI calls.

Also, note that Twitter's DOM is not really huge or deep, which was the worst case for the previous system, and that tests with different count of DOM nodes and different depth, or dirtying DOM elements at the top of the document, could make the improvement way more apparent, while this tests only the hovering over elements almost at the bottom of the DOM.

To summarize, the wins here are two:

 1) Not having to traverse all the node for every tiny incremental restyle, including running flow reconstruction, and
 2) Not having to traverse the whole subtree of an element when marking it as dirty.

I've tried to be fair with the data using a somewhat normal page, but some edge cases could make the servo version without this patch *really* slow.